+++ test-tx.js
@@ -47,7 +47,7 @@ async function testBasic() {
     let tx = new helios.Tx();
 
     tx.addInput(new helios.UTxO(
-        helios.Hash.fromHex("d4b22d33611fb2b3764080cb349b3f12d353aef1d4319ee33e44594bbebe5e83"),
+        helios.TxId.fromHex("d4b22d33611fb2b3764080cb349b3f12d353aef1d4319ee33e44594bbebe5e83"),
         0n,
         new helios.TxOutput(
             helios.Address.fromBech32("addr_test1vzzcg26lxj3twnnx889lrn60pqn0z3km2yahhsz0fvpyxdcj5qp8w"),
@@ -120,7 +120,7 @@ async function testMinting(optimized = false) {
 	let heliosTx = new helios.Tx();
 
 	let mainInput = new helios.UTxO(
-		helios.Hash.fromHex("d4b22d33611fb2b3764080cb349b3f12d353aef1d4319ee33e44594bbebe5e83"),
+		helios.TxId.fromHex("d4b22d33611fb2b3764080cb349b3f12d353aef1d4319ee33e44594bbebe5e83"),
 		0n,
 		new helios.TxOutput(
 			addr,
@@ -185,7 +185,7 @@ async function testInlineDatum() {
 
 	const unsignedBytes = helios.hexToBytes(unsignedHex);
 
-	let inlineDatum = new helios.InlineDatum(new helios.IntData(42n));
+	let inlineDatum = helios.Datum.inline(new helios.IntData(42n));
 
 	console.log(helios.bytesToHex(inlineDatum.toCbor()));